Fix allocation bug in bootloader.
The bootloader was re-using a page for the stack which was in use by
kernel.

This was not detected because the assertion in the vspace didnt
abort when a 4 KiB mapping already exists at a given address (fixed
logic error).

The error happened because of a flaw in the aligned allocation API
of vspace (off by one).

Also added a stack protector page for further hardening and increased
the stack limit.