Fix bug with executor allocation.
We accidentially allocated Ring3Executor from replica 1 but
used them on replica 2. This meant that while it was running
on replica 2, it used the PML4 of replica 1. Which is a disaster.

It kind of works but you can get "spurious pfaults" where
you have to advance the replica, except now advancing means
it really modifies the wrong ptable so it ends up faulting
and faulting till by chance replica 1 gets advanced enough...

This actually manifested in another problem in user-space mutex/cv that
just deadlocked on a cv/mutex pair in rump because it generated so
many pfaults. Not clear at the moment if this is a separate bug
(that manifests because all these pfaults changes lock timings)
or really because we just didn't make any progress at all.