Improve handling of errors while cmdline parsing
Depending on the commandline formatting, there was a case where this
resulted in an endless loop in processing the command line. Errors were
printed using the `error!` macro, that used the logger functionality
before it was setup.

We introduce an error return value for parsing the commandline. It
returns an error whenever there was something wrong with the command
line that could lead to a misconfiguration of the runtime parameters.
For that we have two error values:
 - InvalidCmdLineOptions indicating the command line parsed fine, but
   the selected option are not supported
 - MalformedCmdLine indicating that there were issues with parsing the
   commandline that would have led to skipped arguments.

Signed-off-by: Reto Achermann <achreto@gmail.com>